Allow to place Puput's blog at any sitemap level.
In addition, the same blog slug can be used. The generation of the entries urls and entry_page redirection
have been change to make it possible. Now all the blog url path it is used to find the feeds.

Allow to place the Puput blog at any sitemap level

Change the generation of the entries urls and entry_page redirection to make it possible

feds fixed with same blog slug

improve entries urls and better blog choose on feeds

improve blog search